+++ components/Layout/types.ts
@@ -1,5 +1,4 @@
 type LayoutProps = {
-    title: string;
     lastPublishedAt: string;
     children?: React.ReactNode;
 };
