+++ package.json
@@ -21,9 +21,8 @@
   "dependencies": {
     "humps": "^2.0.1",
     "next": "^14.1.3",
-    "next-themes": "^0.2.1",
+    "next-themes": "^0.4.3",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "sass": "^1.71.1"
+    "react-dom": "^18.2.0"
   }
 }
